Move Profile routes to `route`
