+++ src/main/scala/com/hhandoko/realworld/Server.scala
@@ -13,8 +13,8 @@ import pureconfig.module.catseffect.loadConfigF
 import com.hhandoko.realworld.article.ArticleService
 import com.hhandoko.realworld.auth.{AuthRoutes, AuthService, RequestAuthenticator}
 import com.hhandoko.realworld.config.{Config, DbConfig}
-import com.hhandoko.realworld.profile.{ProfileRoutes, ProfileService}
-import com.hhandoko.realworld.route.ArticleRoutes
+import com.hhandoko.realworld.profile.ProfileService
+import com.hhandoko.realworld.route.{ArticleRoutes, ProfileRoutes}
 import com.hhandoko.realworld.tag.{TagRoutes, TagService}
 import com.hhandoko.realworld.user.{UserRoutes, UserService}
