+++ src/main/scala/com/hhandoko/realworld/route/ProfileRoutes.scala
@@ -1,18 +1,19 @@
-package com.hhandoko.realworld.profile
+package com.hhandoko.realworld.route
 
 import cats.Applicative
 import cats.effect.{ContextShift, Sync}
 import cats.implicits._
 import io.circe.{Encoder, Json}
-import org.http4s.circe.jsonEncoderOf
+import org.http4s.circe._
 import org.http4s.dsl.Http4sDsl
 import org.http4s.{EntityEncoder, HttpRoutes}
 
 import com.hhandoko.realworld.core.Username
+import com.hhandoko.realworld.profile.ProfileService
 
 object ProfileRoutes {
 
-  def apply[F[_]: ContextShift: Sync](profileService: ProfileService[F]): HttpRoutes[F] = {
+  def apply[F[_] : ContextShift : Sync](profileService: ProfileService[F]): HttpRoutes[F] = {
     object dsl extends Http4sDsl[F]; import dsl._
 
     HttpRoutes.of[F] {
@@ -38,8 +39,7 @@ object ProfileRoutes {
       )
     )
 
-    implicit def entityEncoder[F[_]: Applicative]: EntityEncoder[F, ProfileResponse] =
+    implicit def entityEncoder[F[_] : Applicative]: EntityEncoder[F, ProfileResponse] =
       jsonEncoderOf[F, ProfileResponse]
   }
-
 }
